requirements.txt
sounddevice
scipy
librosa
matplotlib
openai
python-dotenv